requirements.txt
asgiref @ file:///opt/conda/conda-bld/asgiref_1661287532135/work
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
Django @ file:///opt/conda/conda-bld/django_1661296381686/work
django-rest-swagger==2.2.0
djangorestframework==3.14.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
openapi-codec==1.3.2
pytz==2022.7.1
requests==2.28.2
simplejson==3.18.3
sqlparse @ file:///croot/sqlparse_1668793487206/work
uritemplate==4.1.1
urllib3==1.26.14
